Split login route from the page
