+++ src/routes/Handler.tsx
@@ -3,7 +3,7 @@ import { BrowserRouter, Switch } from 'react-router-dom'
 import { Public, Private } from './util'
 
 import Dashboard from './pages/dashboard'
-import Login from './pages/login'
+import Login from './pages/login/'
 
 export const Router: React.FC = () => (
   <BrowserRouter>
